requirements.txt
paypal-checkout-serversdk
functions-framework
python-dotenv
pyyaml